# update criterion settings