--- a/algorithms/linfa-clustering/benches/gaussian_mixture.rs
+++ b/algorithms/linfa-clustering/benches/gaussian_mixture.rs
@@ -61,7 +61,7 @@ fn gaussian_mixture_bench(c: &mut Criterion) {
 #[cfg(not(target_os = "windows"))]
 criterion_group! {
   name = benches;
-  config = get_default_profiling_configs();
+  config = config::get_default_profiling_configs();
   targets = gaussian_mixture_bench
 }
 #[cfg(target_os = "windows")]