--- a/algorithms/linfa-ftrl/benches/ftrl.rs
+++ b/algorithms/linfa-ftrl/benches/ftrl.rs
@@ -13,6 +13,13 @@ fn fit_without_prior_model(c: &mut Criterion) {
     let mut rng = SmallRng::seed_from_u64(42);
     let params = Ftrl::params();
     let mut group = c.benchmark_group("Ftrl with no initial model");
+    group
+        .significance_level(0.02)
+        .sample_size(200)
+        .measurement_time(Duration::new(10, 0))
+        .confidence_level(0.97)
+        .warm_up_time(Duration::new(10, 0))
+        .noise_threshold(0.05);
     let sizes: Vec<(usize, usize)> = vec![(10, 1_000), (50, 5_000), (100, 10_000)];
 
     for (nfeatures, nrows) in sizes.iter() {
@@ -34,6 +41,13 @@ fn fit_with_prior_model(c: &mut Criterion) {
     let params = Ftrl::params();
     let valid_params = params.clone().check().unwrap();
     let mut group = c.benchmark_group("Ftrl incremental model training");
+    group
+        .significance_level(0.02)
+        .sample_size(200)
+        .measurement_time(Duration::new(10, 0))
+        .confidence_level(0.97)
+        .warm_up_time(Duration::new(10, 0))
+        .noise_threshold(0.05);
     let sizes: Vec<(usize, usize)> = vec![(10, 1_000), (50, 5_000), (100, 10_000)];
 
     for (nfeatures, nrows) in sizes.iter() {