--- a/src/benchmarks/mod.rs
+++ b/src/benchmarks/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(feature = "benchmarks")]
+pub mod config;