This PR updates the criterion settings for our benched algorithms. Over the course of a day or a few days I will update this description with the timing stats from before and after changes for each algorithm. Unless, it is okay to just do it for a few then I'll do that instead 

### linfa-ica
Old - 1min 39secs
New - 3min 38secs

### linfa-pls
Old - 4mins 16secs
New - 9mins 42secs

### linfa-linear
Old - 1min 29secs
New - 3min 37secs

### linfa-nn
Old - 3mins 43secs
New - 8mins 57secs

### linfa-clustering
#### k-means:
Old - 4min 6secs
New - 7mins 34secs

#### gaussian_mixture
Old - 51secs
New - 2mins 5secs

#### dbscan
Old - 1min 14secs
New - 2mins 50secs

#### appx_db_scan
Old - 47secs
New - 1min 27secs

### linfa-ftrl
Old - 1min 43secs
New - 3min 57secs

### linfa-trees
Old - 55secs
New - 2min 38secs