@YuhanLiin 2022-11-12
Should we refactor the benchmark configs to a common location? Doing it will be a bit involved, since we'll have to add a new `benchmark` feature to `linfa` and put the benchmark configs under that feature.