@oojo12 2022-11-12
Yeah a default would be nice for others and might reduce barriers to contributing benchmarks.