@oojo12 2022-11-13
When I find more time to work on fixing this I will do the above and add a configuration for profiling with pprof. In the spirit of helping out with the investigation related issues.