@YuhanLiin 2022-11-15
Docs are [here](https://doc.rust-lang.org/cargo/reference/features.html). You need to add the feature to the `[features]` section of Cargo.toml and introduce `Criterion` as an optional dependency, then have the new feature include `Critierion`.