@oojo12 2022-11-16
Got the feature in but now there is a peculiar failure for the ubuntu builds regard the tests for linfa-nn. Not sure what's prompting this error as I haven't adjusted the source code. This failure appears on this branch but is not evident when running the same check from wsl on the master branch

`cargo check --workspace --all-targets`