@YuhanLiin 2022-11-16
> `assert_eq!(out, Vec::<ArrayBase<ViewRepr<&f64>, Dim<[usize; 1]>>>::new())` this line fixes the error and seems to work from my local testing. How do you feel about me updating the test?
> 
> Actually, pushing if it isn't desirable, we can always revert.

Try `ArrayView1<&f64>` instead of `ArrayBase<_>`