@oojo12 2022-11-16
> ArrayView1<&f64>

no dice with `Vec::<ArrayView1<&f64>, Dim<[usize; 1]>>::new()` instead of `Vec::<ArrayBase<ViewRepr<&f64>, Dim<[usize; 1]>>>::new()`