@YuhanLiin 2022-11-16
Try `Vec::<ArrayView1<&f64>>`